Make feature converter cls configurable.
Currently, `model.FEATURE_CONVERTER_CLS` is used and it is not configurable.
This means we need to subclass the model class if we just want to use a
different feature converter (e.g., `seqio.PassThroughFeatureConverter` for a
simpler data pipeline).

This CL adds an optional `feature converter_cls` arg and it overrides
`model.FEATURE_CONVERTER_CLS` if provided.

PiperOrigin-RevId: 411473855