[t5x] Don't `None`-out `CheckpointConfig.restore` in `pretrain.gin`.
If `CheckpointConfig.restore` is set to `None`, then the `RestoreCheckpointConfig` will be ignored, which includes any configured `state_transformation_fns`. This is a problem if the pretraining job is stopped and resumed (e.g., if the job is preempted) since `SaveCheckpointConfig` may still contain `state_transformation_fns`, but when the job resumes, the restore-time transformations won't be applied.

PiperOrigin-RevId: 411577854