Add the check that re-writes checkpoints to update between file and gcs readers to the `load_t5x_checkpoint`.
Currently, using the `load_t5x_checkpoint` function on Google Cloud will break if the checkpoint is designed to be read locally. This changes add the check and re-write that depends on where the checkpoint lives (starts with `gs://` or not) that exists in the `restore` method to the `load_t5x_checkpoint` function, allowing for reading of checkpoints on Google Cloud without needing configuration files.

PiperOrigin-RevId: 421127281