Pass a RNG seed into model initalization for eval and infer.
Much of the follow-up work to prompt tuning has required running model evaluations with various different prompts, where some of them are modified after training and don't exist in a checkpoint at all. The current workflow for this is to train a prompt and then run a standalone scripts that merges a prompted checkpoint and a prompt, and then use that result in the evaluation script. This is both time consuming and requires the storing of many copies of almost identical checkpoints which is difficult, especially when the model is large.

This change allows for some variables in a model to be initailized from scratch, even during to evaluation and inference. The Prompt Tuning code base has initializers that allow us to initialize prompts from a file, so this change will let us load a normal t5 checkpoint and a prompt from a file (just like during training), insted of having to merge a special prompted checkpoint with a prompt and then use that.

Allowing fallback_to_scratch to be true is a bit of a footgun, it does make it possible for evalutation to be run with essentially untrained parameters if the user is not careful. Therefore the fallback_init_rng parameter has been default to None, requiring it to be deliberatly set if a user wants to re-initialize variables during eval/infer loading.

PiperOrigin-RevId: 421178494