Update how additional partitioning rules are added.
T5X added the ability to pass a list of partitioning rules into the function that results the default rules for T5-like models. This change udpates the way we get the full set of partitining rules to use this new path instead getting the default rules in our custom partitining rules getter functions. This is the supported way going forward and we do not need to worry about pass arguments into the default T5X rule getter.

PiperOrigin-RevId: 425764414