Ensure the default partitioning rules is a list.
The default logical axis results for 1d sharding of both activations and parameters was a tuple instead of a list. The resulted in an error when additional partitioning rules were provided as the tuple does not have `.extend` method. This changes that one to a list and adds a test that checks that is is possible to add additional rules to all the different possible standard rules settings.

PiperOrigin-RevId: 425818341