Adding SpecialLossNormalizingFactor to T5X
The loss_normalizing_factor (LNF) is currently a float constant which says what to
divide the total summed per-token loss by, to compute the aggregate loss. This
has long been a source of issue in T5X finetuning, since the Adafactor
accumulator magnitude is highly dependent what value it was set to in
pre-training. There are three general recipes that are used:

(1) Set LNF = None during pretraining and finetuning, so the loss is equal to
the total sum decoder log probs for both.
(2) Set LNF = batch_size * decoder_seq_length for pretraining and use the same
value (*not* the finetuning decoder_length) during finetuning.
(3) Set LNF = batch_size * decoder_seq_length for pretraining, and then set it
to finetune_batch_size * finetune_decoder_seq_length for finetuning.

Option (3) is how most DNNs are trained, since it basically just "average log
prob". However, because of padding tokens, this is actually quite difficult to
do correctly in T5X (before this patch). What we really want is to be able to
take the actual average log prob, which is dividing by the number of
*non-padding* target tokens.

This CL adds support for the proper version of (3) by setting:
LNF = SpecialLossNormalizingFactor.NUM_REAL_TARGET_TOKENS
in the Gin config.

It also adds support for a 4th option, which is to set LNF = NUM_REAL_TARGET_TOKENS
during pretraining, but set it to AVERAGE_PER_SEQUENCE during finetuning.

AVERAGE_PER_SEQUENCE still be an average log prob, but will first compute the
average per sequence. This is useful when performing multi-task finetuning
where the tasks have drastically different decoder sequence lengths, but you
still want to weigh them equally.

PiperOrigin-RevId: 426389758