Handle temperature == 0.0 specially in sampled decoding
In sampled decoding, the logits are exponentiated by 1/temperature. To avoid
divide-by-zero issues, we had used an epsilon of 1e-7. However, this number is
too small, resulting in nonsense output when temperature = 0.0 and topk > 1.
Setting a smaller epislon like 1e-4 prevents numeric issues, but does not
always actually result in the greedy output (when topk > 1).

This CL adds a conditional for when temperature = 0.0 (and anything < 1e-4 is
considered 0.0). In this special case, it uses jax.nn.argmax() to select the
top-1 outpu. This will always result in the greedy output, regardless of top-k
or top-p value (which is what the expected behavior is).

PiperOrigin-RevId: 439210746