Supports for 'hard limit' of max_decode_steps in temperature sampling
For the T5X prediction service, the max_decode_steps can be set dynamically by
the user. This can be set to the max input+output size, e.g., 2048. However, in
a shared interactive server, this can incur a very high cost per request.

This CL adds a new parameter to decoding.temperature_sample() called
max_decode_steps_hard_limit which will clip the dynamic max_decode_steps to
this value. This should be set through Gin when the parameter server is launched.

PiperOrigin-RevId: 439430645