[t5x] Sow the encoder input ids and final encoder outputs
This adds sow_intermediate support to the Encoder class to sow the
'final_encoder_outputs'. It also stores 'encoder_input_ids' in the score_batch
intermediates by default. The purpose of this is to extract the final encoder
outputs from the model, without having to sow the activations from every
EncoderLayer, since sowing every encoder layer may result in a large amount of
memory usage for large models.

PiperOrigin-RevId: 452293740