The `logit_callback_fn` of `decoding.temperature_sample` is very useful for implementing custom logits manipulation logic for constrained decoding.
With a host callback, one can even write arbitrary python code without the limitations that JAX imposes.

Currently, the body of the callback has no access to runtime data. This means that we need to recompile whenever the constraints we wish to decode with change. This makes having a single prediction service for requests with different constraints impossible with the current infrastructure.

By adding `logit_callback_arg`, the calling code, for example a prediction service, can pass dynamic run/request time data for the callback logic to use.

One example of how this is useful is implementing a Regular Expression (without lookahead) based constrained decoding `logit_callback_fn`. Logits for tokens that would make the sequence not match the expression can be masked out. This wouldn't be nearly as useful without being able to specify the expression at request time.

`logit_callback_arg` is a string so that it can support as many use cases as possible. More complex data types can easily be serialized into strings by the calling code and parsed in the callback python logic without any of the JAX limitations.

PiperOrigin-RevId: 458552365