[JAX] Fix code that uses buffers after donation on CPU.
JAX has not previously implemented buffer donation on CPU, so donations on CPU had no effect. We are about to implement buffer donation on CPU also, which means if you donate a buffer it may vanish. This change fixes test code that was relying on buffer donation on CPU being a no-op.

PiperOrigin-RevId: 482862439