Log available tasks in increasing editdistance to target task in get_mixture_or_task.
Before this change, when a task name is not found, "Available task:" logs the available tasks in alphabetical order, which seemed to get cut off when the list of available tasks is super long.  This renders this log msg less useful as a source to debug spelling errors in task names, when relevant available task names got cut off.  Adding this additional set of log msgs where available task names are logged in increasing edit distance to target name makes it easier to debug spelling errors according to the log msg.

PiperOrigin-RevId: 487615666