Enable different temperature values for batch items in temperature sampling.
The type signature for the temperature parameter made it seem like the function already supported this but it didn’t.

This change introduces a new code path for when temperature is passed as an array that will branch between `sample_logits_with_nonzero_temperature` and `sample_logits_with_zero_temperature` at the batch item level.

Quickly testing the new and existing code paths on an accelerator shows no degradation of performance.

```
%timeit var_temperature_test(np.random.uniform(size=(8,250000)), jnp.linspace(0, 1, 8))
100 loops, best of 5: 18.1 ms per loop

%timeit var_temperature_test(np.random.uniform(size=(8,250000)), jnp.asarray(0.5))
100 loops, best of 5: 18.2 ms per loop

%timeit var_temperature_test(np.random.uniform(size=(8,250000)), jnp.concatenate((jnp.linspace(0, 1, 5), jnp.array([0,0,0]))))
100 loops, best of 5: 18.6 ms per loop
```

PiperOrigin-RevId: 489236161