[JAX] Remove references to private JAX module jax._src.lib.xla_bridge.
jax._src is a private namespace of JAX, and xla_bridge in particular is going to be moved soon.

Either use jax.lib.xla_bridge which is a more public name, or use .__module__ to identify the correct module without depending on the internal layout of JAX.

In some cases, this change requires JAX from head. We do not promise any form of compatibility to users of jax._src.

PiperOrigin-RevId: 512908226