Fix use with `optax.masked`/`optax.multi_transform`
These transforms produce states that contain MaskedNodes in place of some parameters. This becomes problematic when `OptaxWrapper.derive_logical_axes` constructs an optimizer state object with PartitionSpecs in place of parameters. When `restore_state` gets called with a state dict that has PartitionSpecs as leaves, it will error if it is asked to replace a MaskedNode with a PartitionSpec, because MaskedNodes are NamedTuples. This fix avoids replacing MaskedNodes with PartitionSpecs when building optimizer state objects that have PartitionSpecs in place of parameters.

PiperOrigin-RevId: 544350040