Replace uses of `jnp.array` in types with `jnp.ndarray`.
`jnp.array` is a function, not a type:
https://jax.readthedocs.io/en/latest/_autosummary/jax.numpy.array.html
so it never makes sense to use `jnp.array` in a type annotation. Presumably the intent was to write `jnp.ndarray` aka `jax.Array`.

This CL changes uses of `jnp.array` to `jnp.ndarray`, and adds pytype suppression comments to suppress any type errors that this introduces. This is a trivial and safe change.

PiperOrigin-RevId: 557473961